move to network pool from open-rao-util
