+++ monitoring/pom.xml
@@ -30,6 +30,10 @@
             <artifactId>open-rao-rao-result-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.powsybl</groupId>
+            <artifactId>open-rao-search-tree-rao</artifactId>
+        </dependency>
+        <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>open-rao-util</artifactId>
         </dependency>
@@ -100,11 +104,6 @@
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.powsybl</groupId>
-            <artifactId>open-rao-search-tree-rao</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
