+++ ra-optimisation/search-tree-rao/src/main/java/com/powsybl/openrao/searchtreerao/networkpool/MCDContextWrapper.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.powsybl.openrao.util;
+package com.powsybl.openrao.searchtreerao.networkpool;
 
 import org.slf4j.MDC;
