+++ ra-optimisation/search-tree-rao/src/test/resources/logback-test.xml
@@ -5,11 +5,7 @@
         </encoder>
     </appender>
 
-    <root level="ERROR">
+    <root level="debug">
         <appender-ref ref="STDOUT" />
     </root>
-    <logger name="com.powsybl.openloadflow" level="ERROR"/>
-    <logger name="com.powsybl.math" level="ERROR"/>
-    <logger name="com.powsybl.openrao" level="DEBUG"/>
-    <logger name="com.powsybl.ucte.network.UcteLogger" level="OFF"/>
 </configuration>
